refactor: Refactor TransactionPageErrors for easier overrides
- A helper in errors.js returns all the errors for accounts / triggers

This helper can be overrided to add other errors.

- Errors have a type that is used to choose the component to show

Other components can be used by overriding the errorToComponent attributes
of TransactionPageErrors
