+++ src/ducks/transactions/TransactionPageErrors/index.js
@@ -1,0 +1,3 @@
+import TransactionPageErrors from 'ducks/transactions/TransactionPageErrors/TransactionPageErrors'
+
+export default TransactionPageErrors
