+++ src/ducks/transactions/TriggerErrorCard.spec.jsx
@@ -9,21 +9,24 @@ jest.mock('components/effects', () => ({
 
 describe('trigger error card', () => {
   const setup = () => {
+    const error = {
+      trigger: {
+        current_state: {
+          last_error: 'LOGIN_FAILED'
+        },
+        message: {
+          account: '1234',
+          konnector: 'boursorama83'
+        }
+      }
+    }
     const wrapper = shallow(
       <TestI18n>
         <TriggerErrorCard
           breakpoints={{ isMobile: true }}
           count={1}
           index={0}
-          trigger={{
-            current_state: {
-              last_error: 'LOGIN_FAILED'
-            },
-            message: {
-              account: '1234',
-              konnector: 'boursorama83'
-            }
-          }}
+          error={error}
         />
       </TestI18n>
     )
